Decouple saving from generating
